Camera script
The camera controls have been added, now the camera should follow the mouse whilst in game.

Features a speed of camera movement, and a limiter for the range up and down.

To be used on the camera itself, and has a reference for the player body if needed rotating as well.